pyproject.toml
[tool.poetry]
name = "back-users"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.75.1"
pydantic = {extras = ["email"], version = "^1.9.0"}
pytest = "7.1.1"
uvicorn = "0.17.6"
pytest-cov = "^3.0.0"
python-dotenv = "^0.20.0"
alembic = "^1.7.7"
FastAPI-SQLAlchemy = "^0.2.1"
psycopg2 = "^2.9.3"

[tool.pytest.ini_options]
pythonpath = [
  "app"
]

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
skip-string-normalization = true
